backend/requirements.txt
Django==5.2.7
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-linear-migrations==2.19.0
django-ninja==1.4.5
pydantic==2.12.3
celery==5.5.3
pytest==8.4.2
PyJWT==2.10.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0